feat: add useAuth (28-38)
